fix: update as per comments on PR
